# api-spec.yaml
openapi: "3.0.3"
info:
  title: CareerNexus 求职网站 API
  description: |
    求职网站后端接口文档，供前端开发和 Swagger 调试使用。

    **用户角色**:
    - `JOB_SEEKER` — 求职用户
    - `ENTERPRISE` — 企业用户
    - `ADMIN` — 管理员

    **认证方式**: 除注册和登录外，所有接口需在 Header 中携带 JWT Token:
    ```
    Authorization: Bearer <token>
    ```

    **统一响应格式**: 所有接口均返回 HTTP 200，业务结果由 `code` 字段区分：
    - `code: 1` — 成功，数据在 `data` 字段
    - `code: 0` — 失败，错误信息在 `msg` 字段

    **分页说明**: 分页接口统一使用 `page`（页码，从 1 开始）和 `size`（每页条数）参数，返回 `PaginatedResult`。
  version: "1.0.0"
  contact:
    name: CareerNexus Dev Team
servers:
  - url: http://localhost:8080/api
    description: 开发环境 (直连后端)
  - url: http://localhost/api
    description: nginx 代理环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录后获取的 JWT Token
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        password:
          type: string
          minLength: 6
          maxLength: 32
...
    RegisterJobSeekerRequest:
      type: object
      required: [username, password, email]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        password:
          type: string
          minLength: 6
          maxLength: 32
        email:
          type: string
          format: email
        phone:
          type: string
    RegisterEnterpriseRequest:
      type: object
      required: [username, password, email, companyName]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        password:
          type: string
          minLength: 6
          maxLength: 32
        email:
          type: string
          format: email
        phone:
          type: string
        companyName:
          type: string
...
paths:
...
  /auth/register/job-seeker:
    post:
      tags: [Auth]
      summary: 求职用户注册
      operationId: registerJobSeeker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterJobSeekerRequest'
      responses:
        '200':
          description: |
            注册结果。
            - `code: 1` — 注册成功，`data` 为 User 对象
            - `code: 0` — 用户名或邮箱已存在
  /auth/register/enterprise:
    post:
      tags: [Auth]
      summary: 企业用户注册
      operationId: registerEnterprise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterEnterpriseRequest'
      responses:
        '200':
          description: |
            注册结果。
            - `code: 1` — 注册成功，`data` 为 User 对象
            - `code: 0` — 用户名或邮箱已存在
  /auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: |
            登录结果。
            - `code: 1` — 登录成功，`data` 为 LoginResponse（含 token 和 user）
            - `code: 0` — 用户名或密码错误
...
security:
  - BearerAuth: []